chore: reorganize component folder
