+++ src/app/basket/page.tsx
@@ -3,15 +3,15 @@
 // SPDX-License-Identifier: Apache-2.0
 "use client";
 
-import { useState } from "react";
-import PageHeading from "@/components/PageHeading";
-import DatasetList from "@/components/datasetList";
-import Alert, { AlertState } from "@/components/Alert";
+import Alert, { AlertState } from "@/components/alert";
 import Button from "@/components/button";
-import { faPaperPlane, faPlusCircle } from "@fortawesome/free-solid-svg-icons";
+import PageHeading from "@/components/pageHeading";
 import { useDatasetBasket } from "@/providers/DatasetBasketProvider";
 import { createApplication } from "@/services/daam/index.client";
-import { useSession, signIn } from "next-auth/react";
+import { faPaperPlane, faPlusCircle } from "@fortawesome/free-solid-svg-icons";
+import { signIn, useSession } from "next-auth/react";
+import { useState } from "react";
+import DatasetList from "../datasets/datasetList";
 
 export default function Page() {
   const { basket, isLoading, emptyBasket } = useDatasetBasket();
