+++ src/app/datasets/[id]/page.tsx
@@ -3,10 +3,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import Error from "@/app/error";
-import Chips from "@/components/Chips";
-import PageHeading from "@/components/PageHeading";
-import PageSubHeading from "@/components/PageSubHeading";
-import Sidebar from "@/components/Sidebar";
+import Chips from "@/components/chips";
+import PageHeading from "@/components/pageHeading";
+import PageSubHeading from "@/components/pageSubHeading";
+import Sidebar from "@/components/sidebar";
 import { datasetGet } from "@/services/ckan";
 import AddToBasketBtn from "./AddToBasketBtn";
 import DistributionAccordion from "./DistributionAccordion";
