+++ src/app/layout.tsx
@@ -2,16 +2,16 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import Disclaimer from "@/components/disclaimer";
-import Footer from "@/components/footer";
-import Header from "@/components/header";
-import SessionProviderWrapper from "@/components/utils/sessionProviderWrapper";
 import { DatasetBasketProvider } from "@/providers/DatasetBasketProvider";
 import { config } from "@fortawesome/fontawesome-svg-core";
 import "@fortawesome/fontawesome-svg-core/styles.css";
 import { PublicEnvScript } from "next-runtime-env";
 import { exposeFont, satoshiFont, tabularFont } from "../utils/fonts";
+import Disclaimer from "./disclaimer";
+import Footer from "./footer";
 import "./globals.css";
+import Header from "./header";
+import SessionProviderWrapper from "./sessionProviderWrapper";
 config.autoAddCss = false;
 
 export default function RootLayout({
